Move WifiState into a separate header file
